Update mtl_session02_say.md
Edits to align with See file on mtl.how. On the data UI on my laptop, I only see one scrollbar on the right. In my discussion with Andrew this week, I thought the change to the formatting meant this sentence should be deleted "   a. There are two sets scroll bars on the right. The outer one allows the user to scroll up and down the page, where the inner scroll bar scrolls within the graphs on the page."